backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.25.1
aiohttp==3.9.0
pydantic==2.5.0
